feat: screen에서 padding을 조절하도록 변경

+++ app/src/main/java/nextstep/signup/component/TitleText.kt
@@ -14,7 +14,7 @@ import androidx.compose.ui.unit.sp
 fun TitleText(
     modifier: Modifier = Modifier,
     title: String,
-    fontSize: Float = 26f,
+    fontSize: Int = 26,
     color: Color = Color.Black
 ) {
     Text(
